安卓开发前的调整
将在MSVC中可以通过编译但是gcc不能编译的内容进行修改

+++ LightDiagram/Core/static_traits.h
@@ -759,7 +759,7 @@ public:
 private:
 	field_info(field C:: * _right_ptr, const char* field_name) :
 		field_info(
-			static_cast<size_t>(reinterpret_cast<int>(*(void**)(&_right_ptr))),
+			reinterpret_cast<size_t>(*(void**)(&_right_ptr)),
 			field_name,
 			string(typeid(C).name()) + "::" + field_name,
 			typeid(field)) {}
